Revert "API documentation for Nodes backend"
