+++ desci-server/src/controllers/communities/submissions.ts
@@ -1,11 +1,12 @@
 import { AvailableUserActionLogTypes } from '@desci-labs/desci-models';
 import { ActionType, Submissionstatus } from '@prisma/client';
-import { Response } from 'express';
-import { z } from 'zod';
+import { Request, Response } from 'express';
+import z from 'zod';
 
 import { prisma } from '../../client.js';
-import { BadRequestError, ForbiddenError } from '../../core/ApiError.js';
+import { BadRequestError, ForbiddenError, NotFoundError } from '../../core/ApiError.js';
 import { CreatedSuccessResponse, SuccessMessageResponse, SuccessResponse } from '../../core/ApiResponse.js';
+import { logger } from '../../logger.js';
 import { RequestWithNode, RequestWithUser } from '../../middleware/authorisation.js';
 import {
   createSubmissionSchema,
