+++ desci-server/src/routes/v1/admin/index.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 
 import {
+  analyticsChartSchema,
   createCsv,
   getActiveOrcidUserAnalytics,
   getActiveUserAnalytics,
@@ -15,7 +16,6 @@ import { listAttestations } from '../../../controllers/admin/communities/index.j
 import { debugAllNodesHandler, debugNodeHandler } from '../../../controllers/admin/debug.js';
 import { listDoiRecords, mintDoi } from '../../../controllers/admin/doi/index.js';
 import { resumePublish } from '../../../controllers/admin/publish/resumePublish.js';
-import { analyticsChartSchema } from '../../../controllers/admin/schema.js';
 import { ensureAdmin, ensureUserIsAdmin } from '../../../middleware/ensureAdmin.js';
 import { ensureUser } from '../../../middleware/permissions.js';
 import { validate } from '../../../middleware/validator.js';
