+++ desci-server/src/routes/v1/communities/submissions-schema.ts
@@ -1,70 +1,59 @@
 import { Submissionstatus } from '@prisma/client';
-import 'zod-openapi/extend';
 import { z } from 'zod';
 
 // Schema for creating a new submission
 export const createSubmissionSchema = z.object({
   body: z.object({
     // uuid identifier of the node
-    nodeId: z.string().describe('The unique identifier (UUID) of the research node being submitted'),
-    communityId: z.coerce.number().describe('The ID of the community to submit the research node to'),
+    nodeId: z.string(),
+    communityId: z.coerce.number(),
   }),
 });
 
 // Schema for getting community submissions
 export const getCommunitySubmissionsSchema = z.object({
   params: z.object({
-    communityId: z.coerce.number().describe('The ID of the community to get submissions from'),
+    communityId: z.coerce.number(),
   }),
   query: z.object({
-    status: z
-      .enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED])
-      .optional()
-      .describe('Filter submissions by their status (pending, accepted, or rejected)'),
+    status: z.enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED]).optional(),
   }),
 });
 
 // Schema for getting user submissions
 export const getUserSubmissionsSchema = z.object({
   params: z.object({
-    userId: z.string().describe('The ID of the user to get submissions for'),
+    userId: z.string(),
   }),
   query: z.object({
-    status: z
-      .enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED])
-      .optional()
-      .describe('Filter submissions by their status (pending, accepted, or rejected)'),
+    status: z.enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED]).optional(),
   }),
 });
 
 // Schema for updating submission status
 export const updateSubmissionStatusSchema = z.object({
   params: z.object({
-    submissionId: z.coerce.number().describe('The ID of the submission to update'),
+    submissionId: z.coerce.number(),
   }),
   body: z.object({
-    status: z
-      .enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED])
-      .describe('The new status to set for the submission'),
+    status: z.enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED]),
   }),
 });
 
 // Schema for getting a single submission
 export const getSubmissionSchema = z.object({
   params: z.object({
-    submissionId: z.coerce.number().describe('The ID of the submission to retrieve'),
+    submissionId: z.coerce.number(),
   }),
 });
 
-// Schema for rejecting a submission
+// Schema for getting a single submission
 export const rejectSubmissionSchema = z.object({
   params: z.object({
-    submissionId: z.coerce.number().describe('The ID of the submission to reject'),
+    submissionId: z.coerce.number(),
   }),
   body: z.object({
-    reason: z.string().optional().describe('Optional reason for rejecting the submission'),
-    status: z
-      .enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED])
-      .describe('The status to set for the submission (typically REJECTED)'),
+    reason: z.string().optional(),
+    status: z.enum([Submissionstatus.PENDING, Submissionstatus.ACCEPTED, Submissionstatus.REJECTED]),
   }),
 });
