+++ desci-server/src/server.ts
@@ -14,13 +14,11 @@ import helmet from 'helmet';
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import { pinoHttp } from 'pino-http';
 import { Server as SocketIOServer } from 'socket.io';
-import swaggerUI from 'swagger-ui-express';
 import { v4 } from 'uuid';
 
 import { orcidConnect } from './controllers/auth/orcid.js';
 import { orcidCheck } from './controllers/auth/orcidNext.js';
 import { NotFoundError } from './core/ApiError.js';
-import { openaiDocumentation } from './docs/openai.js';
 import { als, logger as parentLogger } from './logger.js';
 import { RequestWithUser } from './middleware/authorisation.js';
 import { ensureUserIfPresent } from './middleware/ensureUserIfPresent.js';
@@ -28,9 +26,8 @@ import { errorHandler } from './middleware/errorHandler.js';
 import { extractAuthToken, extractUserFromToken } from './middleware/permissions.js';
 import routes from './routes/index.js';
 import { initializeWebSockets, getIO } from './services/websocketService.js';
-import swaggerFile from './swagger_output.json' with { type: 'json' };
 import { SubmissionQueueJob } from './workers/doiSubmissionQueue.js';
-// import { runWorkerUntilStopped } from './workers/publish.js';
+import { runWorkerUntilStopped } from './workers/publish.js';
 
 // const __dirname = path.dirname(__filename);
 
@@ -213,8 +210,6 @@ class AppServer {
     });
     this.app.get('/orcid', orcidConnect);
     this.app.post('/orcid/next', [ensureUserIfPresent], orcidCheck());
-    this.app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerFile));
-    this.app.use('/documentation', swaggerUI.serve, swaggerUI.setup(openaiDocumentation));
     this.app.use('/', routes);
   }
