--- desci-server/src/swagger.ts
@@ -1,6 +1,0 @@
-import swaggerAutogen from 'swagger-autogen';
-
-const outputFile = './swagger_output.json';
-const endpointsFiles = ['./routes/index.ts'];
-
-swaggerAutogen()(outputFile, endpointsFiles);
